requirements.txt
pygit2>=1.14.0
sv-ttk>=2.6.0
Pillow>=10.0.0
tray-manager>=0.1.0
darkdetect[macos-listener]>=0.8.0
psutil>=5.9.0
dbus-next>=0.2.3
async-tkinter-loop>=0.10.3
# Build tools (optional, for packaging)
pyinstaller>=6.0.0; extra == "build"